Incorporate review changes
